backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
chromadb==0.4.18
sentence-transformers==2.2.2
pymupdf==1.23.8
python-docx==1.1.0
google-generativeai==0.3.2
tiktoken==0.5.1
pydantic==2.5.0
PyPDF2==3.0.1
pillow==10.4.0
pytesseract==0.3.10
langchain==0.2.11
langchain-community==0.2.10
